// src/types/mock-data.type.ts
export type MockData = {
    titles: string[];
    descriptions: string[];
    genres: string[];
    previews: string[];
    videos: string[];
    actors: string[];
    directors: string[];
    names: string[];
    emails: string[];
    avatars: string[];
    passwords: string[];
    posters: string[];
    backgroundImages: string[];
    backgroundColors: string[];
};